Ну, почти все

+++ app/src/main/java/com/example/kolodeznapp/API/JokeApi.kt
@@ -11,7 +11,7 @@ import com.example.kolodeznapp.model.Joke
  */
 interface JokeApi {
 
-    fun deleteJoke(id: Int)
+    fun deleteJoke(id: Long)
 
     fun editJoke(joke: Joke)
